refactor(frontend): align service imports and fix lint warnings
