+++ app/frontend/src/composables/generation/useGenerationStudio.ts
@@ -1,8 +1,10 @@
 import { computed, onMounted } from 'vue'
 
-import { useGenerationPersistence } from '@/composables/generation'
-import { useGenerationStudioController } from '@/composables/generation/useGenerationStudioController'
-import { useGenerationUI } from '@/composables/generation'
+import {
+  useGenerationPersistence,
+  useGenerationStudioController,
+  useGenerationUI,
+} from '@/composables/generation'
 import { useDialogService, useNotifications } from '@/composables/shared'
 import { useGenerationFormStore } from '@/stores/generation'
 import type { GenerationFormState, NotificationType } from '@/types'
