+++ app/frontend/src/composables/generation/useJobQueueTransport.ts
@@ -1,6 +1,6 @@
 import { ref, unref, type MaybeRefOrGetter, type Ref } from 'vue';
 
-import { fetchActiveGenerationJobs } from '@/services/generation/generationService';
+import { fetchActiveGenerationJobs } from '@/services';
 import { DEFAULT_BACKEND_BASE } from '@/utils/backend';
 import type { GenerationJobStatus } from '@/types';
 
@@ -43,6 +43,9 @@ export const useJobQueueTransport = (
       return records;
     } catch (error) {
       apiAvailable.value = false;
+      if (import.meta.env.DEV) {
+        console.error('[JobQueue] Failed to fetch generation jobs', error);
+      }
       return null;
     }
   };
