+++ app/frontend/src/composables/system/useSystemStatus.ts
@@ -1,8 +1,7 @@
 import { computed } from 'vue';
 import { storeToRefs } from 'pinia';
 
-import { useGenerationConnectionStore } from '@/stores/generation';
-import { useSystemStatusController } from '@/stores/generation/systemStatusController';
+import { useGenerationConnectionStore, useSystemStatusController } from '@/stores/generation';
 
 const formatMemory = (used: number, total: number) => {
   if (!total) {
